This patch contains the whole Act 2 from Spencer to Colemak, completly outlined, with maps.
It also haves a quest for Act 1 which was done by DroidBotAI. Delay.
Title files.

It have some infodumps and is very compresed, so the whole Act 2 can be
done in less than 5 minutes, 10 atop.
However all you need, from going to 2nd continent, having basic
knowledge, knowing a little the cryonic RR Resorts (you mess and you go
there -- see IRCResolutions on wiki), to uncryonizing Colemak is done
with minimal difficulty.
The two NPC dying is also unexplained and makes no sense (just wow), it
should display "-- Continues --" big on top but it's not displayed
currently. Just making players anxious with content which is only
outined on the wiki and as Matthias requested, there is no access to it.