revive_faction(): THIS IS A HACK AND SHOULD BE REVERTED BEFORE RELEASE
This commit is aimed solely for testers and must be reverted before
merging this branch on master.
It includes a lua function revive_faction() which revives MS bots.

It must be removed for a couple of reasons:
Improper coding with hardcoded-limits,
Bots on Act2 must be from 'rr' faction, or not affected by kill_faction()

Kindly revert this commit when it becomes anymore necessary.
Thank you.