Fix hidden bugs on win_game() at the cost of a new ending.
Until this commit, Colemak finished the game.
However win_game() doesn't cleans cursor, so if you clicked to go to
Colemak, the cursor click could be preserved and credits skipped.

To fix this bug, the ending was moved to the Terminal.
This is done with a new command, 'download', to get Dvorak's instructions.
Part from instructions were copied from titles/EndOfGame.lua
Maybe we should play a title, instead of using dialog interface?

After issuing the download command there'll be a text message from Colemak.
Dvorak will complain about not being able to hear you, and about missing
Qwerty and Arensito (names not mentioned). In the end, instead of saying
player to go to the Ice Pass, he'll instruct player to go to Contribute
section or to wait a new release.

For reference purposes, after the game is end this time without bugs, the
game version will be displayed.