Continue work on area south of RR Resorts.
This includes a new NPC, Master Yadda, which is only a draft currently.
The gate is now named but it'll stay closed.
The C-64 virus hub is on World Edge. Tux must use a secret passage on
Icy Summer Artificial Island which is closed for obvious reasons.

Icy Summer Artificial Island was built south of RR Resorts for enjoyment from
staff. It's called “Icy Summer” because it tries to reproduce summer in a land
where ice is the rule.

As described in Comments, Master Yadda will also teach Tux the Light skill,
give tux the Light Saber and Light Staff too, besides selling Laser Power Pack
and Small Plasma Grenades. But this all, only in the future.