Works on Master Yadda to the work can be submitted.
You still won't have access to Light Sabers and other cool features.
But Master Yadda will unlock the gate for the C64 for you.

The requirement is talk to him 37 times. Good luck with that.
While this unlocks the C-64 hub, it's by no means complete.
It should - and will - be improved to be more than killing a bot,
with lots of HP and DMG.

The area south of RR Resorts is still locked.