Some important changes as requested by Infrared to improve Act 2 quality.
This is a follow-up. It introduces the usage of Quest Log, adds blood
on RR Resorts which will be deleted after a while playing the game, but which
explains a lot about the Arcane Lore, moves a dead body nearer to it,
modify the transition sequence so you get some basic explanation of what is
going on, and minor, less relevant changes to some dialog files.