Prevent player from getting reward from Iris twice, minor wiki corrections.
What should be 'purpose' was marked as 'backstory', this has been corrected,
and a backstory field was added (although it may require more discussion from
developers).

The core of this commit consists in checking that the player still have the
Desk Lamp when trading with Iris. (Deleting an inexistent item does NOT throws
an error in FreedroidRPG, not even a console warning, apparently)