Consistency: Modify Ewald and Francis' Chest, concerning mugs and cups.
In modern life, when you go to a Bar, they supply you a cup to drink, and then
required it returned. How to explain that in FreedroidRPG you must bring your
own mug? Is the bar really that sunk?

This patch removes the Mug from Francis' Chest, and gives it to player when
talking to Ewald on FirstTime. Ewald then warns you to bring it any time you
want something to drink. It may be a campaign against discardable items usage,
which produced a lot of trash - probably it started decades before, and was not
discontinued.

As proposed enhancement, you could receive the full mess kit with Ewald, or
someone else (eg. Francis) could have this dialog with Tux - in that case,
they would explain that everyone have their own dish, and Tux will need his own
too in case he wants to eat anything.