Consistency: Change some dialogs involving Bender, East Gate and Lily password.
Formerly Bender used to guard east gate after getting cured, however, as time
went by and developers changed the game, Bender returned to his starting
point at doctor's room entrance.

This lead to an inconsistent dialog, in which Bender claims to be in guard
duty of a gate with a 614 - which isn't true anymore. This fixes this dialog,
by removing references to the gate and the 614.

It also changes c-net terminal, because Lily's password is an asterix repeated
four times according to Bender, but you wrote six digits and got logged in.
Now if you know her password you'll insert four asterix. (Consistency!)