Minor redesign in Master Arena, based at level 59.
This redesign grants the Master Arena some traps and more walls.
It is not difficult at all, if you could clear the Elite Arena.

It may need to be rebalanced/redone (maybe entirely) later.
As usual, if you manage to fulfill the Master Arena:
you probably are ready for the Hell Fortress.