Better keyboard input handling
We still have some issues with the keyboard, especially when a modifier key is pressed.

We currently use one single function (getchar_raw()) for 2 purposes:
1) get chars to fill an editable string input widget (mainly in the leveleditor)
2) get key presses to fire in-game (or in-lvleditor) actions (those defined in the keychart)

In the first use case, we need 'ascii' values, in the second case, we need 'raw' values.
getchar_raw() was trying to do its best to answer to those 2 needs, but failed...

This patch defines 2 separate functions: getchar_ascii() and getchar_raw(), specific to each use case.

Related to R2362
Closes #235 and #765