Allow to cancel ongoing attack with a right-click.
Currently, the only way to stop an ongoing attack once a range-weapon
is 'locked' to a target is to click on a non-occupied spot, but that
initiates walking.

With this patch, right-clicking on an anemy stops the current automatic
shoot, and activate the selected program.

Based on a patch proposed by Robin Gareus.

Relates to RR2368