i18n: Add some 'xgettext:no-c-format' comment where needed.
By default, xgettext uses some heuristics to find if an i18n string
is a C format string. It then adds a 'c-format' comment before the msgid
string in the .po file.

When msgfmt later compiles the l10n catalog (.mo file), it checks if the
translated string is still a valid format string.

But sometimes, the heuristics are wrong.
For instance: "This is 100% natural" is detected as a C format (due to
'% n'). In some language, once 'natural' is translated, it results in
an invalid C format string...

To avoid this, a 'xgettext:no-c-format' comment MUST be added before
the i18n string.