Do not redirect the console output, if already redirected to a file or to a pipe (Unix only)
When the game detects that it is not run from a terminal, stdout and stderr are
redirected to .freedroid_rpg/fdrpg_out.txt.

If the output was already redirected to a file or a pipe by the user, the game
thinks that it is not run from a terminal, and thus it redirects its output.
The user's command is then not fulfilled.

This patch detects such a situation to disable the game's own redirection.