When saving an item, write the itemspec's id rather than the item index in ItemMap.
If an item is added in the middle of the item_specs file, the index is
indeed no more the right one.
In the best case, a wrong item will be loaded.
In the worst case, an item of a wrong type (source book instead of weapon,
for instance) is loaded, and the game can crash when using it.

Add a savegame converter, which apply the conversion on old savegames.
-> new savegame revision is 001601-08