Include Energy Shield skill, but only as an alternative to Invisibility.
You're forced to rely on Energy Shield for "not taking damage", and this skill
becomes outdated when Invisibility starts being sold (before HF fight at library,
or at some special quests)

Because Invisibility actually makes you Invulnerable (except against boss), and
even allows you to attack while on this state, this skill does not follow the
original purpose.

In future, what we currently know as Invisibility will become an Invulnerability
skill, and the Invisibility skill will not allow you to attack while it is active.

Relates to RR1984
Closes #511