Do not attempt to teleport a dead NPC.
If a lua script tries to teleport a dead NPC, the game craches because
the npc is in no level's list.
This patch informs the player about the bug, and avoids the teleportation.

This closes issue923

But note that the actual fix would be to check if a npc is alive before
to teleport it, in lua scripts.