Some item's inventory images stick out of their slot in the item list.
This happens when there is too much difference between the image's ratio
and the window's ratio. The scale applied to the image is wrongly
computed, resulting in a too wide or too tall displayed image.

This patch scales correctly the displayed image to fit in its slot,
by constraining the width or the height depending on the ratio of the
image AND the ratio of the window.

This patch also factorizes the computations for OpenGL and SDL modes,
to ensure identical results.