Restore the current font at the end of the display of a menu.
Currently, if the font is changed in an option of the menu, this font
becomes the new current font.
The next frame rendering will then render the whole menu using that
'current font'.

This patch stores the current font before to display the menu, and
restores it after the menu is displayed.