savegame: fix game loading on non-glibc platforms
savegame loading uses %ms scanf format to match a string and allocate it, but this is only available on some platforms.
The "m" modifier is apparently in POSIX, but it's not available in Haiku, and scanf there will not do anything for %ms.

Fix the bug by manually allocating memory and reading into it.