Rework commit dade2a74, which was reverted by 0aa3161c.
The reworked commit was about "preventing a crash, possibly, due to FP
arithmetic accuracy problems..."

Here is the comment added in view.c:resolve_virtual_position():

We need to ensure that the computed position is inside the boundaries of
the 'real' level. Due to a lack of precision during floating point computations
(i.e. if x<<<y, then (y+x) => (y)), we could possibly get a wrong output.
For instance, if vpos.x = -0.000002, we expect the computed position to be
rpos.x = 89.999998 on the left level (given a level's width of 90).
But, using floats, 90.0-0.000002 gives 90.0, which is not "inside" the level,
and thus this function returns false.
The same issue applies to small positive quantities.
As a fix, a "too small" value is replaced by EPSILON_FIX, which is high enough
to be representable in the range of float values that we use for the positions.
Note that this limits the position's precision, around level's boundary, to 1/1000,
which seems to be small enough.

Closes issue917