Fix a bug on droid animations, in SDL mode, with precaching enabled.
Currently the animation images are preloaded when a droid spec is parsed from
droid_specs.lua.
The droid specs are stored in a dynarray. So, when the dynarray is reallocated,
the pointers associated to the images become invalid, and animations are
broken.

This patch moves the preload of enemy graphics at a later stage, once
the entire set of droid specs is loaded.

This closes issue964.