Fix display of accented characters on MacOSX
Since MacOSX 10.14 (Mojave), the system locale can no more been retrieved
with a call to setlocale(LC_MESSAGES, ""), when run by double-clicking on the
App's icon.
It however works as expected when run from the command line...

If the setlocale() call does not succeed, the system language is now
requested using the CoreFoundation API.

This fixes issue954.