No need to warn if a config parameter is missing in fdrpg.cfg
If a parameter is missing, its default value is used, so it is useless
to warn the user.

This patch extends the gen_savestruct.py script, so that some structures
can be read without warnings (see the 'no_warn_for_structs' new array).

The 'configuration_for_freedroid' structure is one of those (the only
one actually).

Fixes issue976