Heap overflow in loading untrusted save game
Fixed length buffers are used to decode the maps and the waypoints
when loading a savegame. Corrupted savegames could then be used to
create an heap overflow and execute malicious code.

This path fixes 2 reported issues :

1/ The length of the map's line currently extracted from the savegame
is computed, and used to allocated a storage buffer. The same is done
for decoding a waypoint.

The length of the savegame's line currently read is also checked and if
it is too long, the game is aborted.

This fixes issue951.

2/ When decoding a map, the expected line length is computed (based on
the map xlen and the number of layers) and if the line read from the
savegame does not have that expected length, the game is aborted.

This fixes issue952.