Use a dynarray to store the items of a level.
Currently, level's items are stored in a static C array.
If the array is full, then it is no more possible to drop an item
to the floor.
Moreover, when swapping items in the inventory, the item hold in hand
is temporary stored in the level's items array. And if that array
is full, the game crashes.

This patch replaces the static C array by a dynarray, thus removing
any limit on the number of items that a level can contain.

This fixes issue966