Change the way data dirs are looked up.
When run from an AppImage, the rootfs is mounted under a temporary
subdir defined by an APPDIR envvar.

Absolute data dirs, provided by the FD_DATADIR and LOCALEDIR
build defines, must then be prefixed by the APPDIR.

Moreover, when creating the AppImage, datas are stored into the
FD_DATADIR, and so relative datadir paths are useless.