Fix 'function declaration isn't a prototype' warnings
This adds void to function declarations with empty parameter lists in
order to remove warnings found by -Wstrict-prototypes. The functions in
getopt.{c,h} were not changed because it appears to be deliberate there
(so that the prototypes are compatible with the prototypes found on
systems that don't use glibc).

Closes issue2411