Fix variable shadowing warnings
This fixes the warnings found by -Wshadow.

In most cases, a local variable was called the same name as a type name,
so the local variable was renamed. There were some cases in which a
local variable was declared again inside a loop (for instance).
In these cases, the local variable was also renamed.

Closes issue2412