Fix crash in TutorialTom
If the Tutorial is saved before to request to continue to the second
section, and the savegame is reloaded, then the game crashes when we
ask to continue to the next section.

This is due to a Lua array being set in the FirstTime script, and
reused in a later node's script.
Lua arrays are not (yet) saved, and so the array is unknown when it is
reused.

The array is actually static, so it can be define as a 'chunk local'
variable to fix the crash.