Fix potential offseting of the minigame
When the takeover minigame is started, the User_Rect is reset to
use the whole screen, so that all graphics are correctly centered.

But, if the takeover minigame is opened by a dialog script, then the
User_Rect is not reset. Consequently, if a panel (skill/inv/stat) was
opened before the minigame starts, then graphics are offset.

This patch replaces the use of User_Rect by the use of Full_User_Rect